requirements.txt
asgiref==3.2.5
Django==3.0.4
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
